pyproject.toml
[tool.poetry]
name = "quotes"
version = "0.1.0"
description = ""
authors = ["Aman Rawat <ar.rawataman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.7"
django-autoslug = "^1.9.9"
django-tailwind = "^3.6.0"
shortuuid = "^1.0.11"
pillow = "^10.1.0"
gunicorn = "^21.2.0"
django-htmx = "^1.17.2"
dj-database-url = "^2.1.0"
psycopg = "^3.1.18"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
faker = "^20.1.0"
django-debug-toolbar = "^4.2.0"
django-browser-reload = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
